refactor(tailwind): move menus to tailwind
