--- src/components/Menu/style.module.css
@@ -1,163 +1,0 @@
-/*
-License:
-
-This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of
-the MPL was not distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
-*/
-
-.menu_button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 4rem;
-  list-style-type: none;
-  background-color: transparent;
-}
-.menu_button::-webkit-details-marker {
-  display: none;
-}
-.menu_button:is(:hover, :active) {
-  background-color: #16a34a;
-  background-color: var(--header-hover-bg-color);
-}
-.menu_button:focus {
-  box-shadow:
-    0 0 0 3px rgba(59, 130, 246, 0.5),
-    0 1px 3px 0 rgba(0, 0, 0, 0.1),
-    0 1px 2px 0 rgba(0, 0, 0, 0.06);
-  box-shadow:
-    0 0 0 3px highlight,
-    0 1px 3px 0 rgba(0, 0, 0, 0.1),
-    0 1px 2px 0 rgba(0, 0, 0, 0.06);
-  outline: 2px solid transparent;
-  outline-offset: 2px;
-}
-.menu_button_img {
-  filter: invert(100%);
-}
-
-.container {
-  position: absolute;
-  top: 0px;
-  top: env(safe-area-inset-top, 0px);
-  right: 0px;
-  right: env(safe-area-inset-right, 0px);
-  margin-top: 3rem;
-  max-height: 90%;
-  max-height: calc(100vh - 3rem);
-  max-height: calc(
-    100vh - 3rem - env(safe-area-inset-top, 0px) -
-      env(safe-area-inset-bottom, 0px)
-  );
-  overflow-y: auto;
-  overscroll-behavior-y: contain;
-  color: white;
-  color: var(--header-color);
-  background-color: #064e3b;
-  background-color: var(--header-bg-color);
-  box-shadow:
-    0 10px 15px -3px rgba(0, 0, 0, 0.1),
-    0 4px 6px -2px rgba(0, 0, 0, 0.05);
-}
-
-.scroll_container {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  justify-content: center;
-  padding: 0.75rem;
-}
-
-.element_container {
-  display: flex;
-  flex-direction: column;
-}
-
-.label {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  margin-top: 1.25rem;
-  text-align: center;
-  color: white;
-  color: var(--header-color);
-}
-
-.month_full_year_toggle {
-  composes: button from "/src/form.module.css";
-  padding-top: 0.5rem;
-  padding-bottom: 0.5rem;
-  margin-top: 1.25rem;
-}
-
-.month_select {
-  composes: button from "/src/form.module.css";
-  height: 2.5rem;
-}
-
-.form_item {
-  composes: button from "/src/form.module.css";
-  flex: 1 1 auto;
-  margin-top: 0.25rem;
-  height: 2.5rem;
-}
-
-.groups {
-  composes: form_item;
-  margin-top: 1.25rem;
-}
-
-.small_button {
-  composes: button from "/src/form.module.css";
-  height: 3rem;
-  padding: 0.5rem 1rem;
-  margin-top: 1.25rem;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.share_address {
-  padding: 0.25rem;
-  margin-left: -0.25rem;
-  background-color: transparent;
-  border: 0px;
-  border-radius: 0.25rem;
-  color: white;
-  color: var(--header-color);
-}
-.share_address:focus {
-  box-shadow:
-    0 0 0 3px rgba(59, 130, 246, 0.5),
-    0 1px 3px 0 rgba(0, 0, 0, 0.1),
-    0 1px 2px 0 rgba(0, 0, 0, 0.06);
-  box-shadow:
-    0 0 0 3px highlight,
-    0 1px 3px 0 rgba(0, 0, 0, 0.1),
-    0 1px 2px 0 rgba(0, 0, 0, 0.06);
-  outline: 2px solid transparent;
-  outline-offset: 2px;
-}
-
-.title {
-  padding: 0px;
-  margin: 0px;
-  margin-top: 1.25rem;
-  font-size: 1.125rem;
-  line-height: 1.75rem;
-}
-
-.share_label {
-  margin-top: 1.25rem;
-}
-
-.buttons_row {
-  display: flex;
-  flex-direction: row;
-  margin-top: 1.25rem;
-}
-.buttons_row > :is(button, a) {
-  flex: 1 1 8rem;
-}
-.buttons_row > *:last-child {
-  margin-left: 1.5rem;
-}
